+++ runtime/executor/src/main/java/org/apache/nemo/runtime/executor/datatransfer/BlockOutputWriter.java
@@ -1,0 +1,105 @@
+/*
+ * Copyright (C) 2018 Seoul National University
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.nemo.runtime.executor.datatransfer;
+
+import org.apache.nemo.common.ir.edge.executionproperty.*;
+import org.apache.nemo.common.ir.vertex.IRVertex;
+import org.apache.nemo.runtime.common.RuntimeIdManager;
+import org.apache.nemo.runtime.common.plan.RuntimeEdge;
+import org.apache.nemo.runtime.executor.data.BlockManagerWorker;
+import org.apache.nemo.runtime.executor.data.block.Block;
+import org.apache.nemo.runtime.executor.data.partitioner.*;
+
+import java.util.Map;
+import java.util.Optional;
+
+/**
+ * Represents the output data transfer from a task.
+ */
+public final class BlockOutputWriter extends OutputWriter {
+  private final DataStoreProperty.Value blockStoreValue;
+  private final BlockManagerWorker blockManagerWorker;
+  private final Block blockToWrite;
+  private final boolean nonDummyBlock;
+
+  /**
+   * Constructor.
+   *
+   * @param hashRangeMultiplier the {@link org.apache.nemo.conf.JobConf.HashRangeMultiplier}.
+   * @param srcTaskId           the id of the source task.
+   * @param dstIrVertex         the destination IR vertex.
+   * @param runtimeEdge         the {@link RuntimeEdge}.
+   * @param blockManagerWorker  the {@link BlockManagerWorker}.
+   */
+  BlockOutputWriter(final int hashRangeMultiplier,
+                    final String srcTaskId,
+                    final IRVertex dstIrVertex,
+                    final RuntimeEdge<?> runtimeEdge,
+                    final BlockManagerWorker blockManagerWorker) {
+    super(hashRangeMultiplier, dstIrVertex, runtimeEdge);
+    this.blockManagerWorker = blockManagerWorker;
+    this.blockStoreValue = runtimeEdge.getPropertyValue(DataStoreProperty.class)
+      .orElseThrow(() -> new RuntimeException("No data store property on the edge"));
+    blockToWrite = blockManagerWorker.createBlock(
+        RuntimeIdManager.generateBlockId(getId(), srcTaskId), blockStoreValue);
+    final Optional<DuplicateEdgeGroupPropertyValue> duplicateDataProperty =
+        runtimeEdge.getPropertyValue(DuplicateEdgeGroupProperty.class);
+    nonDummyBlock = !duplicateDataProperty.isPresent()
+        || duplicateDataProperty.get().getRepresentativeEdgeId().equals(runtimeEdge.getId())
+        || duplicateDataProperty.get().getGroupSize() <= 1;
+  }
+
+  /**
+   * Writes output element depending on the communication pattern of the edge.
+   *
+   * @param element the element to write.
+   */
+  public void write(final Object element) {
+    if (nonDummyBlock) {
+      blockToWrite.write(partitioner.partition(element), element);
+
+      final DedicatedKeyPerElement dedicatedKeyPerElement =
+          partitioner.getClass().getAnnotation(DedicatedKeyPerElement.class);
+      if (dedicatedKeyPerElement != null) {
+        blockToWrite.commitPartitions();
+      }
+    } // If else, does not need to write because the data is duplicated.
+  }
+
+  /**
+   * Notifies that all writes for a block is end.
+   * Further write about a committed block will throw an exception.
+   */
+  public void close() {
+    // Commit block.
+    final DataPersistenceProperty.Value persistence =
+        runtimeEdge.getPropertyValue(DataPersistenceProperty.class).
+            orElseThrow(() -> new RuntimeException("No data persistence property on the edge"));
+
+    final Optional<Map<Integer, Long>> partitionSizeMap = blockToWrite.commit();
+    // Return the total size of the committed block.
+    if (partitionSizeMap.isPresent()) {
+      long blockSizeTotal = 0;
+      for (final long partitionSize : partitionSizeMap.get().values()) {
+        blockSizeTotal += partitionSize;
+      }
+      this.writtenBytes = blockSizeTotal;
+    } else {
+      this.writtenBytes = -1; // no written bytes info.
+    }
+    blockManagerWorker.writeBlock(blockToWrite, blockStoreValue, getExpectedRead(), persistence);
+  }
+}
